Added a .gitignore file
The beginning of this file contains the 'standard' ignores for all the
go build files, but also includes the executables that can/are built by
the 'all' target, so those are not included.

Those could be excluded in cmd/buffer and cmd/hello instead, but I chose
to do them here so I have a single place to consult.